client_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
django==4.2.27
djangorestframework==3.15.2
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.1.4
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.5
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
google-auth-oauthlib==1.2.1
